only genotype to add

+++ src/main/java/agh/ics/oop/IEngineObserver.java
@@ -1,5 +1,5 @@
 package agh.ics.oop;
 
 public interface IEngineObserver {
-    public void stepMade(Integer epoch, Integer grasNumber, Integer animalsNumber,float avgEnergy, float avgChildrenNum, float avgLifeLength);
+    public void stepMade(Integer epoch, Float grasNumber, Float animalsNumber,Float avgEnergy, Float avgChildrenNum, Float avgLifeLength);
 }
