+++ src/main/java/agh/ics/oop/gui/TabelMaintainer.java
@@ -70,6 +70,6 @@ public class TabelMaintainer implements IEngineObserver {
     }
 
     @Override
-    public void stepMade(Integer epoch, Integer grasNumber, Integer animalsNumber, float avgEnergy, float avgChildrenNum, float avgLifeLength) {updateTable();}
+    public void stepMade(Integer epoch, Float grasNumber, Float animalsNumber, Float avgEnergy, Float avgChildrenNum, Float avgLifeLength) {updateTable();}
 }
